# Improve performance of GEMM for small matrices when SMP is defined.