@martin-frbg 2018-06-07
Interesting find. Would you happen to have some numbers for before/after, just to save me some cpu time ?