@oon3m0oo 2018-06-08
Sure, no problem. There's a bit of variance in the individual numbers in part because of the uncertainty of scheduling, but you get the idea.  The numbers are most consistent for the smaller sizes.  What's interesting is that enabling, but not using, SMP has a real cost for small matrices.  I've typically seen 15-20ns slowdowns.  With this change that's halved.

Without OpenMP:
Run on (12 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 15360K (x1)
-----------------------------------------------------
Benchmark              Time           CPU Iterations
-----------------------------------------------------
BM_SGEMM/4           103 ns        103 ns   40555859
BM_SGEMM/6           175 ns        175 ns   24406039
BM_SGEMM/8           201 ns        201 ns   20917720
BM_SGEMM/10          267 ns        267 ns   15713849
BM_SGEMM/16          469 ns        469 ns    9016428
BM_SGEMM/20          647 ns        647 ns    6565700
BM_SGEMM/32         1904 ns       1904 ns    2169187
BM_SGEMM/40         2958 ns       2958 ns    1415388
BM_SGEMM/64         9313 ns       9314 ns     452245
BM_SGEMM/80        15539 ns      15539 ns     265024
BM_SGEMM/100       29698 ns      29699 ns     140856
BM_SGEMM/128       56055 ns      56056 ns      76945
BM_SGEMM/150       99996 ns      99998 ns      42131
BM_SGEMM/200      202282 ns     202285 ns      20724
BM_SGEMM/256      403832 ns     403840 ns      10442
BM_SGEMM/300      672956 ns     672958 ns       6297
BM_SGEMM/400     1472337 ns    1472362 ns       2873
BM_SGEMM/500     2928214 ns    2928281 ns       1470
BM_SGEMM/600     4887383 ns    4887490 ns        838
BM_SGEMM/700     7915470 ns    7915597 ns        536
BM_SGEMM/800    11728328 ns   11728457 ns        359
BM_SGEMM/1000   22711137 ns   22711507 ns        186
BM_SGEMM/2000  186067446 ns  186072020 ns         23
BM_DGEMM/4           104 ns        104 ns   40923660
BM_DGEMM/6           143 ns        143 ns   29552295
BM_DGEMM/8           178 ns        178 ns   23552126
BM_DGEMM/10          272 ns        272 ns   15365230
BM_DGEMM/16          528 ns        528 ns    7974881
BM_DGEMM/20          906 ns        906 ns    4568169
BM_DGEMM/32         2759 ns       2759 ns    1563884
BM_DGEMM/40         5495 ns       5495 ns     766055
BM_DGEMM/64        16171 ns      16171 ns     260008
BM_DGEMM/80        30154 ns      30154 ns     139502
BM_DGEMM/100       55007 ns      55009 ns      77416
BM_DGEMM/128      108331 ns     108334 ns      39241
BM_DGEMM/150      178880 ns     178883 ns      23171
BM_DGEMM/200      370712 ns     370720 ns      11612
BM_DGEMM/256      760480 ns     760505 ns       5540
BM_DGEMM/300     1205314 ns    1205321 ns       3523
BM_DGEMM/400     2722566 ns    2722642 ns       1555
BM_DGEMM/500     5287194 ns    5287284 ns        814
BM_DGEMM/600     9019084 ns    9019265 ns        466
BM_DGEMM/700    14372062 ns   14372438 ns        297
BM_DGEMM/800    21142685 ns   21142953 ns        198
BM_DGEMM/1000   40974767 ns   40975114 ns        103
BM_DGEMM/2000  323346596 ns  323354365 ns         13

With OpenMP, without this change (note the proportionally higher times for small matrices):
Run on (12 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 15360K (x1)
-----------------------------------------------------
Benchmark              Time           CPU Iterations
-----------------------------------------------------
BM_SGEMM/4           118 ns        118 ns   35649367
BM_SGEMM/6           199 ns        199 ns   21961462
BM_SGEMM/8           220 ns        220 ns   19117809
BM_SGEMM/10          290 ns        290 ns   14276746
BM_SGEMM/16          493 ns        493 ns    8493607
BM_SGEMM/20          662 ns        662 ns    6336222
BM_SGEMM/32         1920 ns       1920 ns    2188407
BM_SGEMM/40         2987 ns       2987 ns    1413241
BM_SGEMM/64         9457 ns       9457 ns     448854
BM_SGEMM/80        29773 ns      29685 ns     139096
BM_SGEMM/100       32844 ns      31912 ns     100000
BM_SGEMM/128       59445 ns      59439 ns      68872
BM_SGEMM/150       64148 ns      64142 ns      55778
BM_SGEMM/200       79921 ns      79921 ns      54385
BM_SGEMM/256      213205 ns     204189 ns      21964
BM_SGEMM/300      227416 ns     227417 ns      18147
BM_SGEMM/400      485271 ns     485277 ns       9317
BM_SGEMM/500      903494 ns     903473 ns       4881
BM_SGEMM/600     1449618 ns    1416460 ns       3084
BM_SGEMM/700     2564857 ns    2415026 ns       1787
BM_SGEMM/800     3776359 ns    3463524 ns       1273
BM_SGEMM/1000    5848950 ns    5800890 ns        763
BM_SGEMM/2000   40336905 ns   40205035 ns        103
BM_DGEMM/4           118 ns        118 ns   35700879
BM_DGEMM/6           158 ns        158 ns   26805544
BM_DGEMM/8           198 ns        198 ns   21225266
BM_DGEMM/10          299 ns        299 ns   14092777
BM_DGEMM/16          553 ns        553 ns    7639389
BM_DGEMM/20          926 ns        926 ns    4563911
BM_DGEMM/32         2770 ns       2769 ns    1534507
BM_DGEMM/40         4821 ns       4821 ns     874363
BM_DGEMM/64        16094 ns      16094 ns     262146
BM_DGEMM/80        25063 ns      25062 ns     168676
BM_DGEMM/100       28542 ns      28402 ns     146210
BM_DGEMM/128       60996 ns      59852 ns      77607
BM_DGEMM/150       80078 ns      79801 ns      59220
BM_DGEMM/200      146820 ns     135038 ns      39475
BM_DGEMM/256      330169 ns     306777 ns      12117
BM_DGEMM/300      342618 ns     336485 ns      10289
BM_DGEMM/400      753562 ns     738728 ns       6142
BM_DGEMM/500     1383266 ns    1349299 ns       3039
BM_DGEMM/600     2351780 ns    2299085 ns       1996
BM_DGEMM/700     4012654 ns    4000423 ns       1132
BM_DGEMM/800     5233760 ns    5221856 ns        780
BM_DGEMM/1000   10518069 ns   10107650 ns        422
BM_DGEMM/2000   70905081 ns   70665583 ns         54

With OpenMP, with this change:
Run on (12 X 3800 MHz CPU s)
CPU Caches:
  L1 Data 32K (x6)
  L1 Instruction 32K (x6)
  L2 Unified 256K (x6)
  L3 Unified 15360K (x1)
-----------------------------------------------------
Benchmark              Time           CPU Iterations
-----------------------------------------------------
BM_SGEMM/4           110 ns        110 ns   38099136
BM_SGEMM/6           184 ns        184 ns   23227383
BM_SGEMM/8           209 ns        209 ns   20159236
BM_SGEMM/10          270 ns        270 ns   15301951
BM_SGEMM/16          476 ns        476 ns    8732949
BM_SGEMM/20          646 ns        646 ns    6482065
BM_SGEMM/32         1905 ns       1906 ns    2214764
BM_SGEMM/40         2959 ns       2959 ns    1416487
BM_SGEMM/64         9356 ns       9356 ns     450680
BM_SGEMM/80        30483 ns      30473 ns     135416
BM_SGEMM/100       23630 ns      23630 ns     178687
BM_SGEMM/128       61182 ns      61182 ns      69124
BM_SGEMM/150       65658 ns      65523 ns      63477
BM_SGEMM/200       81440 ns      81025 ns      37781
BM_SGEMM/256      196737 ns     196731 ns      21946
BM_SGEMM/300      226863 ns     226855 ns      18959
BM_SGEMM/400      469335 ns     462479 ns       9142
BM_SGEMM/500      847270 ns     801694 ns       5285
BM_SGEMM/600     1272040 ns    1270250 ns       3339
BM_SGEMM/700     2394877 ns    2362031 ns       1806
BM_SGEMM/800     3125238 ns    3121107 ns       1000
BM_SGEMM/1000    5828526 ns    5552087 ns        778
BM_SGEMM/2000   42210184 ns   42017687 ns         96
BM_DGEMM/4           106 ns        106 ns   39474718
BM_DGEMM/6           145 ns        145 ns   28700178
BM_DGEMM/8           186 ns        186 ns   22514105
BM_DGEMM/10          281 ns        281 ns   14905388
BM_DGEMM/16          536 ns        536 ns    7855043
BM_DGEMM/20          912 ns        912 ns    4603354
BM_DGEMM/32         2750 ns       2750 ns    1544001
BM_DGEMM/40         4836 ns       4836 ns     857973
BM_DGEMM/64        16261 ns      16261 ns     260024
BM_DGEMM/80        25568 ns      25567 ns     142644
BM_DGEMM/100       29889 ns      29775 ns     149308
BM_DGEMM/128       55217 ns      55216 ns      78212
BM_DGEMM/150       72809 ns      72809 ns      54463
BM_DGEMM/200      109297 ns     109271 ns      36067
BM_DGEMM/256      238937 ns     238933 ns      17511
BM_DGEMM/300      339181 ns     335320 ns      12389
BM_DGEMM/400      671850 ns     626274 ns       6754
BM_DGEMM/500     1764091 ns    1521852 ns       3299
BM_DGEMM/600     2344661 ns    2126935 ns       1891
BM_DGEMM/700     3615566 ns    3595405 ns       1152
BM_DGEMM/800     5264189 ns    5249601 ns        802
BM_DGEMM/1000   10700431 ns   10493148 ns        429
BM_DGEMM/2000   71360596 ns   70966913 ns         61